Fixed issue with initializing external API
